fix: Исправлены ошибки в работе сервиса docs: Доработан файл README
