+++ README.md
@@ -1,1 +1,22 @@
-# service-host-discovery
+<h1 align="center" style="border-bottom: none !important; margin-bottom: 5px !important;"><a href="#">Service Host Discovery</a></h1>
+<p align="center">Сервис для обнаружения хостов в сети</p>
+<p align="center">Host Detection Service on Network</p>
+<p align="center">
+  <a href="#">
+    <img src="https://img.shields.io/github/license/crocup/Service-Host-Discovery" />
+  </a>
+<a href="#">
+    <img src="https://img.shields.io/github/last-commit/crocup/Service-Host-Discovery" />
+  </a>
+  <a href="#">
+    <img src="https://img.shields.io/github/stars/crocup/Service-Host-Discovery?style=social" />
+  </a>
+</p>
+
+# About
+Сервис предназначен для обнаружения хостов в сети.
+Данный сервис используется в ПО Fenix Security Scanner
+
+## Warning
+Сервис находится в разработке!
+<p>This is a development version, we don't recommend its use in production!</p>
