+++ test.py
@@ -1,0 +1,9 @@
+import unittest
+
+
+class TestTask(unittest.TestCase):
+    pass
+
+
+if __name__ == '__main__':
+    unittest.main()
